+++ src/lib/InputInternal.tsx
@@ -3,6 +3,7 @@ import { Input } from "reactstrap";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { InputProps } from "./Input";
 import { useMarkOnFocusHandler } from "./hooks/useMarkOnFocusHandler";
+import { useInternalFormContext } from "./context/InternalFormContext";
 
 // This is two random guids concatenated. It is used to set the value of the option to undefined.
 const UNDEFINED_OPTION_VALUE = "CABB7A27DB754DA58C89D43ADB03FE0EC5EE3E25A6624D749F35CF2E92CFA784";
@@ -38,6 +39,7 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
 
   const fieldError = get(errors, name) as FieldError | undefined;
   const hasError = !!fieldError;
+  const { disabled: formDisabled } = useInternalFormContext<T>();
 
   return (
     <>
@@ -50,7 +52,7 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
         max={rangeMax}
         rows={textAreaRows}
         multiple={multiple}
-        disabled={disabled}
+        disabled={formDisabled || disabled}
         plaintext={plainText}
         style={plainText ? { color: "black", marginLeft: 10, ...style } : { ...style }}
         placeholder={placeholder}