+++ src/lib/context/InternalFormContext.tsx
@@ -3,23 +3,27 @@ import { FieldValues } from "react-hook-form";
 
 export interface InternalFormContextProps<T extends FieldValues> {
   requiredFields: (keyof T)[];
+  disabled: boolean;
 }
 
-export interface InternalFormContextProviderProps<T extends FieldValues> extends Pick<InternalFormContextProps<T>, "requiredFields"> {
+export interface InternalFormContextProviderProps<T extends FieldValues>
+  extends Pick<InternalFormContextProps<T>, "requiredFields" | "disabled"> {
   children: React.ReactNode;
 }
 
 export const InternalFormContext = createContext<InternalFormContextProps<never>>({
   requiredFields: [],
+  disabled: false,
 });
 
 export const InternalFormProvider = <T extends FieldValues>(props: InternalFormContextProviderProps<T>) => {
-  const { children, requiredFields } = props;
+  const { children, requiredFields, disabled } = props;
 
   return (
     <InternalFormContext.Provider
       value={{
         requiredFields,
+        disabled,
       }}
     >
       {children}