+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support to have undefined values in a select input by using a magic guid value
 
+## [1.5.0] - 2023-09-18
+
+### Added
+
+- Added `disabled` property to disable all inputs in the form
+
+### Changed
+
+- Exports `useInternalFormContext` and its properties through children function params
+
 ### dependabot: \#71 Bump tj-actions/changed-files from 38 to 39
 
 ### dependabot: \#65 Bump tj-actions/changed-files from 37 to 38
@@ -296,7 +306,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/1.4.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/1.5.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -321,6 +331,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[1.5.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.3.3...1.4.0
 [1.3.3]: https://github.com/neolution-ch/react-hook-form-components/compare/1.3.2...1.3.3
 [1.3.2]: https://github.com/neolution-ch/react-hook-form-components/compare/1.3.1...1.3.2